Module aws_sdk_textract::error

Modules

access_denied_exception

See AccessDeniedException

bad_document_exception

See BadDocumentException

document_too_large_exception

See DocumentTooLargeException

human_loop_quota_exceeded_exception

See HumanLoopQuotaExceededException

idempotent_parameter_mismatch_exception

See IdempotentParameterMismatchException

internal_server_error

See InternalServerError

invalid_job_id_exception

See InvalidJobIdException

invalid_kms_key_exception

See InvalidKmsKeyException

invalid_parameter_exception

See InvalidParameterException

invalid_s3_object_exception

See InvalidS3ObjectException

limit_exceeded_exception

See LimitExceededException

provisioned_throughput_exceeded_exception

See ProvisionedThroughputExceededException

throttling_exception

See ThrottlingException

unsupported_document_exception

See UnsupportedDocumentException

Structs

AccessDeniedException

You aren't authorized to perform the action. Use the Amazon Resource Name (ARN)

AnalyzeDocumentError

AnalyzeExpenseError

BadDocumentException

Amazon Textract isn't able to read the document. For more information on the document

DetectDocumentTextError

DocumentTooLargeException

The document can't be processed because it's too large. The maximum document size for

GetDocumentAnalysisError

GetDocumentTextDetectionError

HumanLoopQuotaExceededException

Indicates you have exceeded the maximum number of active human in the loop workflows available

IdempotentParameterMismatchException

A ClientRequestToken input parameter was reused with an operation, but at

InternalServerError

Amazon Textract experienced a service issue. Try your call again.

InvalidJobIdException

An invalid job identifier was passed to GetDocumentAnalysis or to

InvalidKmsKeyException

Indicates you do not have decrypt permissions with the KMS key entered, or the KMS key

InvalidParameterException

An input parameter violated a constraint. For example, in synchronous operations,

InvalidS3ObjectException

Amazon Textract is unable to access the S3 object that's specified in the request.

LimitExceededException

An Amazon Textract service limit was exceeded. For example, if you start too many

ProvisionedThroughputExceededException

The number of requests exceeded your throughput limit. If you want to increase this limit,

StartDocumentAnalysisError

StartDocumentTextDetectionError

ThrottlingException

Amazon Textract is temporarily unable to process the request. Try your call again.

UnsupportedDocumentException

The format of the input document isn't supported. Documents for synchronous operations can be in

Enums

AnalyzeDocumentErrorKind

AnalyzeExpenseErrorKind

DetectDocumentTextErrorKind

GetDocumentAnalysisErrorKind

GetDocumentTextDetectionErrorKind

StartDocumentAnalysisErrorKind

StartDocumentTextDetectionErrorKind