Module aws_sdk_textract::model

Modules

block

See Block

bounding_box

See BoundingBox

document

See Document

document_location

See DocumentLocation

document_metadata

See DocumentMetadata

expense_detection

See ExpenseDetection

expense_document

See ExpenseDocument

expense_field

See ExpenseField

expense_type

See ExpenseType

geometry

See Geometry

human_loop_activation_output

See HumanLoopActivationOutput

human_loop_config

See HumanLoopConfig

human_loop_data_attributes

See HumanLoopDataAttributes

line_item_fields

See LineItemFields

line_item_group

See LineItemGroup

notification_channel

See NotificationChannel

output_config

See OutputConfig

point

See Point

relationship

See Relationship

s3_object

See S3Object

warning

See Warning

Structs

Block

A Block represents items that are recognized in a document within a group

BoundingBox

The bounding box around the detected page, text, key-value pair, table, table cell, or

Document

The input document, either as bytes or as an S3 object.

DocumentLocation

The Amazon S3 bucket that contains the document to be processed. It's used by asynchronous

DocumentMetadata

Information about the input document.

ExpenseDetection

An object used to store information about the Value or Label detected by Amazon Textract.

ExpenseDocument

The structure holding all the information returned by AnalyzeExpense

ExpenseField

Breakdown of detected information, seperated into

ExpenseType

An object used to store information about the Type detected by Amazon Textract.

Geometry

Information about where the following items are located on a document page: detected

HumanLoopActivationOutput

Shows the results of the human in the loop evaluation. If there is no HumanLoopArn, the

HumanLoopConfig

Sets up the human review workflow the document will be sent to if one of the conditions

HumanLoopDataAttributes

Allows you to set attributes of the image. Currently, you can declare an image as free

LineItemFields

A structure that holds information about the different lines found in a document's tables.

LineItemGroup

A grouping of tables which contain LineItems, with each table identified by the table's LineItemGroupIndex.

NotificationChannel

The Amazon Simple Notification Service (Amazon SNS) topic to which Amazon Textract publishes the completion status of

OutputConfig

Sets whether or not your output will go to a user created bucket. Used to set the name

Point

The X and Y coordinates of a point on a document page. The X and Y

Relationship

Information about how blocks are related to each other. A Block object

S3Object

The S3 bucket name and file name that identifies the document.

Warning

A warning about an issue that occurred during asynchronous text analysis (StartDocumentAnalysis) or asynchronous document text detection (StartDocumentTextDetection).

Enums

BlockType

ContentClassifier

EntityType

FeatureType

JobStatus

RelationshipType

SelectionStatus

TextType