Struct aws_sdk_textract::model::block::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Block

Implementations

impl Builder

pub fn block_type(self, input: BlockType) -> Self

The type of text item that's recognized. In operations for text detection, the following types are returned:

PAGE - Contains a list of the LINE Block objects that are detected on a document page.

WORD - A word detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

In text analysis operations, the following types are returned:

PAGE - Contains a list of child Block objects that are detected on a document page.

KEY_VALUE_SET - Stores the KEY and VALUE Block objects for linked text that's detected on a document page. Use the EntityType field to determine if a KEY_VALUE_SET object is a KEY Block object or a VALUE Block object.

WORD - A word that's detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

TABLE - A table that's detected on a document page. A table is grid-based information with two or more rows or columns, with a cell span of one row and one column each.

CELL - A cell within a detected table. The cell is the parent of the block that contains the text in the cell.

SELECTION_ELEMENT - A selection element such as an option button (radio button) or a check box that's detected on a document page. Use the value of SelectionStatus to determine the status of the selection element.

pub fn set_block_type(self, input: Option<BlockType>) -> Self

pub fn confidence(self, input: f32) -> Self

The confidence score that Amazon Textract has in the accuracy of the recognized text and the accuracy of the geometry points around the recognized text.

pub fn set_confidence(self, input: Option<f32>) -> Self

pub fn text(self, input: impl Into<String>) -> Self

The word or line of text that's recognized by Amazon Textract.

pub fn set_text(self, input: Option<String>) -> Self

pub fn text_type(self, input: TextType) -> Self

The kind of text that Amazon Textract has detected. Can check for handwritten text and printed text.

pub fn set_text_type(self, input: Option<TextType>) -> Self

pub fn row_index(self, input: i32) -> Self

The row in which a table cell is located. The first row position is 1. RowIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

pub fn set_row_index(self, input: Option<i32>) -> Self

pub fn column_index(self, input: i32) -> Self

The column in which a table cell appears. The first column position is 1. ColumnIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

pub fn set_column_index(self, input: Option<i32>) -> Self

pub fn row_span(self, input: i32) -> Self

The number of rows that a table cell spans. Currently this value is always 1, even if the number of rows spanned is greater than 1. RowSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

pub fn set_row_span(self, input: Option<i32>) -> Self

pub fn column_span(self, input: i32) -> Self

The number of columns that a table cell spans. Currently this value is always 1, even if the number of columns spanned is greater than 1. ColumnSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

pub fn set_column_span(self, input: Option<i32>) -> Self

pub fn geometry(self, input: Geometry) -> Self

The location of the recognized text on the image. It includes an axis-aligned, coarse bounding box that surrounds the text, and a finer-grain polygon for more accurate spatial information.

pub fn set_geometry(self, input: Option<Geometry>) -> Self

pub fn id(self, input: impl Into<String>) -> Self

The identifier for the recognized text. The identifier is only unique for a single operation.

pub fn set_id(self, input: Option<String>) -> Self

pub fn relationships(self, input: impl Into<Relationship>) -> Self

pub fn set_relationships(self, input: Option<Vec<Relationship>>) -> Self

pub fn entity_types(self, input: impl Into<EntityType>) -> Self

pub fn set_entity_types(self, input: Option<Vec<EntityType>>) -> Self

pub fn selection_status(self, input: SelectionStatus) -> Self

The selection status of a selection element, such as an option button or check box.

pub fn set_selection_status(self, input: Option<SelectionStatus>) -> Self

pub fn page(self, input: i32) -> Self

The page on which a block was detected. Page is returned by asynchronous operations. Page values greater than 1 are only returned for multipage documents that are in PDF format. A scanned image (JPEG/PNG), even if it contains multiple document pages, is considered to be a single-page document. The value of Page is always 1. Synchronous operations don't return Page because every input document is considered to be a single-page document.

pub fn set_page(self, input: Option<i32>) -> Self

pub fn build(self) -> Block

Consumes the builder and constructs a Block