Struct aws_sdk_textract::model::Block

#[non_exhaustive]
pub struct Block {

    pub block_type: Option<BlockType>,
    pub confidence: Option<f32>,
    pub text: Option<String>,
    pub text_type: Option<TextType>,
    pub row_index: Option<i32>,
    pub column_index: Option<i32>,
    pub row_span: Option<i32>,
    pub column_span: Option<i32>,
    pub geometry: Option<Geometry>,
    pub id: Option<String>,
    pub relationships: Option<Vec<Relationship>>,
    pub entity_types: Option<Vec<EntityType>>,
    pub selection_status: Option<SelectionStatus>,
    pub page: Option<i32>,

}

A Block represents items that are recognized in a document within a group of pixels close to each other. The information returned in a Block object depends on the type of operation. In text detection for documents (for example DetectDocumentText), you get information about the detected words and lines of text. In text analysis (for example AnalyzeDocument), you can also get information about the fields, tables, and selection elements that are detected in the document.

An array of Block objects is returned by both synchronous and asynchronous operations. In synchronous operations, such as DetectDocumentText, the array of Block objects is the entire set of results. In asynchronous operations, such as GetDocumentAnalysis, the array is returned over one or more responses.

For more information, see How Amazon Textract Works.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

block_type: Option<BlockType>

The type of text item that's recognized. In operations for text detection, the following types are returned:

PAGE - Contains a list of the LINE Block objects that are detected on a document page.

WORD - A word detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

In text analysis operations, the following types are returned:

PAGE - Contains a list of child Block objects that are detected on a document page.

KEY_VALUE_SET - Stores the KEY and VALUE Block objects for linked text that's detected on a document page. Use the EntityType field to determine if a KEY_VALUE_SET object is a KEY Block object or a VALUE Block object.

WORD - A word that's detected on a document page. A word is one or more ISO basic Latin script characters that aren't separated by spaces.

LINE - A string of tab-delimited, contiguous words that are detected on a document page.

TABLE - A table that's detected on a document page. A table is grid-based information with two or more rows or columns, with a cell span of one row and one column each.

CELL - A cell within a detected table. The cell is the parent of the block that contains the text in the cell.

SELECTION_ELEMENT - A selection element such as an option button (radio button) or a check box that's detected on a document page. Use the value of SelectionStatus to determine the status of the selection element.

confidence: Option<f32>

The confidence score that Amazon Textract has in the accuracy of the recognized text and the accuracy of the geometry points around the recognized text.

text: Option<String>

The word or line of text that's recognized by Amazon Textract.

text_type: Option<TextType>

The kind of text that Amazon Textract has detected. Can check for handwritten text and printed text.

row_index: Option<i32>

The row in which a table cell is located. The first row position is 1. RowIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

column_index: Option<i32>

The column in which a table cell appears. The first column position is 1. ColumnIndex isn't returned by DetectDocumentText and GetDocumentTextDetection.

row_span: Option<i32>

The number of rows that a table cell spans. Currently this value is always 1, even if the number of rows spanned is greater than 1. RowSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

column_span: Option<i32>

The number of columns that a table cell spans. Currently this value is always 1, even if the number of columns spanned is greater than 1. ColumnSpan isn't returned by DetectDocumentText and GetDocumentTextDetection.

geometry: Option<Geometry>

The location of the recognized text on the image. It includes an axis-aligned, coarse bounding box that surrounds the text, and a finer-grain polygon for more accurate spatial information.

id: Option<String>

The identifier for the recognized text. The identifier is only unique for a single operation.

relationships: Option<Vec<Relationship>>

A list of child blocks of the current block. For example, a LINE object has child blocks for each WORD block that's part of the line of text. There aren't Relationship objects in the list for relationships that don't exist, such as when the current block has no child blocks. The list size can be the following:

0 - The block has no child blocks.

1 - The block has child blocks.

entity_types: Option<Vec<EntityType>>

The type of entity. The following can be returned:

KEY - An identifier for a field on the document.

VALUE - The field text.

EntityTypes isn't returned by DetectDocumentText and GetDocumentTextDetection.

selection_status: Option<SelectionStatus>

The selection status of a selection element, such as an option button or check box.

page: Option<i32>

The page on which a block was detected. Page is returned by asynchronous operations. Page values greater than 1 are only returned for multipage documents that are in PDF format. A scanned image (JPEG/PNG), even if it contains multiple document pages, is considered to be a single-page document. The value of Page is always 1. Synchronous operations don't return Page because every input document is considered to be a single-page document.

Implementations

impl Block

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Block